// src/app/tramitacion/[foliolicencia]/[idoperador]/c1/(modelos)/formulario-type.ts
export interface formularioApp {
  accion: 'siguiente' | 'guardar' | 'navegar';
  linkNavegacion: string;
  run: string;
  razon: string;
  telefono: string;
  fecharecepcionlme: string;
  region: string;
  comuna: string;
  tipo: string;
  fechaemision: string;
  calle: string;
  numero: string;
  departamento: string;
  actividadlaboral: string;
  ocupacion: string;
  otro: string;
}